+++ CHANGELOG.md
@@ -69,6 +69,7 @@ a huge list of updates and fixes.
 - SN no longer accepts objects with invocation or verification script bigger than 1KiB (#3887)
 - Every object with non-zero payload is now paid, not only regular ones (#3856)
 - Separate policer placement state from replica shortage (#3901)
+- SN now forwards remote SN's response to the client as is (#3877)
 
 ### Removed
 - `node.persistent_sessions.path` config option from SN config (#3846)